[#7] modify files name, icon props
